feat: core extended

+++ framework/streaming/chat.go
@@ -88,7 +88,7 @@ func (a *Accumulator) processAccumulatedChatStreamingChunks(requestID string, re
 	if accumulator.StartTimestamp.IsZero() || accumulator.FinalTimestamp.IsZero() {
 		data.Latency = 0
 	} else {
-		data.Latency = float64(accumulator.FinalTimestamp.Sub(accumulator.StartTimestamp).Nanoseconds()) / 1e6
+		data.Latency = accumulator.FinalTimestamp.Sub(accumulator.StartTimestamp).Nanoseconds() / 1e6
 	}
 	data.EndTimestamp = accumulator.FinalTimestamp
 	data.OutputMessage = completeMessage
